EES-2687 - broke out ReleaseApprovalService and dependencies from ReleaseService in order to reduce complexity of ReleaseService and better encapsulate work around approval status updates in a central area
