+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/ReleaseService.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces;
-using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.ManageContent;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.Security;
 using GovUk.Education.ExploreEducationStatistics.Admin.ViewModels;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
@@ -23,12 +22,9 @@ using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
 using static GovUk.Education.ExploreEducationStatistics.Admin.Validators.ValidationErrorMessages;
 using static GovUk.Education.ExploreEducationStatistics.Admin.Validators.ValidationUtils;
-using static GovUk.Education.ExploreEducationStatistics.Common.Model.FileType;
 using static GovUk.Education.ExploreEducationStatistics.Content.Model.MethodologyPublishingStrategy;
 using static GovUk.Education.ExploreEducationStatistics.Content.Model.MethodologyStatus;
 using IReleaseRepository = GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.IReleaseRepository;
-using IReleaseService = GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces.IReleaseService;
-using Publication = GovUk.Education.ExploreEducationStatistics.Content.Model.Publication;
 using Release = GovUk.Education.ExploreEducationStatistics.Content.Model.Release;
 using Unit = GovUk.Education.ExploreEducationStatistics.Common.Model.Unit;
 
@@ -38,7 +34,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
     {
         private readonly ContentDbContext _context;
         private readonly StatisticsDbContext _statisticsDbContext;
-        private readonly IPublishingService _publishingService;
         private readonly IMapper _mapper;
         private readonly IPersistenceHelper<ContentDbContext> _persistenceHelper;
         private readonly IUserService _userService;
@@ -50,18 +45,14 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
         private readonly IDataImportService _dataImportService;
 	    private readonly IFootnoteService _footnoteService;
         private readonly IDataBlockService _dataBlockService;
-        private readonly IReleaseChecklistService _releaseChecklistService;
-        private readonly IContentService _contentService;
         private readonly IReleaseSubjectRepository _releaseSubjectRepository;
         private readonly IGuidGenerator _guidGenerator;
-        private readonly IPreReleaseUserService _preReleaseUserService;
 
         // TODO EES-212 - ReleaseService needs breaking into smaller services as it feels like it is now doing too
         // much work and has too many dependencies
         public ReleaseService(
             ContentDbContext context,
             IMapper mapper,
-            IPublishingService publishingService,
             IPersistenceHelper<ContentDbContext> persistenceHelper,
             IUserService userService,
             IReleaseRepository repository,
@@ -73,14 +64,10 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
             IFootnoteService footnoteService,
             StatisticsDbContext statisticsDbContext,
             IDataBlockService dataBlockService,
-            IReleaseChecklistService releaseChecklistService,
-            IContentService contentService,
             IReleaseSubjectRepository releaseSubjectRepository,
-            IGuidGenerator guidGenerator,
-            IPreReleaseUserService preReleaseUserService)
+            IGuidGenerator guidGenerator)
         {
             _context = context;
-            _publishingService = publishingService;
             _mapper = mapper;
             _persistenceHelper = persistenceHelper;
             _userService = userService;
@@ -93,11 +80,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
             _footnoteService = footnoteService;
             _statisticsDbContext = statisticsDbContext;
             _dataBlockService = dataBlockService;
-            _releaseChecklistService = releaseChecklistService;
-            _contentService = contentService;
             _releaseSubjectRepository = releaseSubjectRepository;
             _guidGenerator = guidGenerator;
-            _preReleaseUserService = preReleaseUserService;
         }
 
         public async Task<Either<ActionResult, ReleaseViewModel>> GetRelease(Guid id)
@@ -106,31 +90,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
                 .CheckEntityExists<Release>(id, HydrateReleaseForReleaseViewModel)
                 .OnSuccess(_userService.CheckCanViewRelease)
                 .OnSuccess(release => _mapper.Map<ReleaseViewModel>(release));
-        }
-
-        // TODO EES-2687 - extract Release Status methods into separate Service  
-        public async Task<Either<ActionResult, List<ReleaseStatusViewModel>>> GetReleaseStatuses(
-            Guid releaseId)
-        {
-            return await _persistenceHelper
-                .CheckEntityExists<Release>(releaseId, release =>
-                    release.Include(r => r.ReleaseStatuses)
-                        .ThenInclude(rs => rs.CreatedBy))
-                .OnSuccess(_userService.CheckCanViewReleaseStatusHistory)
-                .OnSuccess(release =>
-                    release.ReleaseStatuses
-                        .Select(rs =>
-                            new ReleaseStatusViewModel
-                            {
-                                ReleaseStatusId = rs.Id,
-                                InternalReleaseNote = rs.InternalReleaseNote,
-                                ApprovalStatus = rs.ApprovalStatus,
-                                Created = rs.Created,
-                                CreatedByEmail = rs.CreatedBy?.Email
-                            })
-                        .OrderByDescending(vm => vm.Created)
-                        .ToList()
-                );
         }
 
         public async Task<Either<ActionResult, ReleaseViewModel>> CreateRelease(ReleaseCreateViewModel releaseCreate)
@@ -326,7 +285,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
                 .OnSuccess(_mapper.Map<ReleasePublicationStatusViewModel>);
         }
 
-        // TODO EES-2687 - extract Release Status methods into separate Service  
         public async Task<Either<ActionResult, ReleaseViewModel>> UpdateRelease(
             Guid releaseId, ReleaseUpdateViewModel request)
         {
@@ -346,102 +304,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
                     await _context.SaveChangesAsync();
                     return await GetRelease(releaseId);
                 });
-        }
-
-        // TODO EES-2687 - extract Release Status methods into separate Service  
-        public async Task<Either<ActionResult, ReleaseViewModel>> CreateReleaseStatus(
-            Guid releaseId, ReleaseStatusCreateViewModel request)
-        {
-            return await _persistenceHelper
-                .CheckEntityExists<Release>(releaseId, ReleaseChecklistService.HydrateReleaseForChecklist)
-                .OnSuccess(_userService.CheckCanUpdateRelease)
-                .OnSuccessDo(release => _userService.CheckCanUpdateReleaseStatus(release, request.ApprovalStatus))
-                .OnSuccess(async release =>
-                {
-                    if (request.ApprovalStatus != ReleaseApprovalStatus.Approved && release.Published.HasValue)
-                    {
-                        return ValidationActionResult(PublishedReleaseCannotBeUnapproved);
-                    }
-
-                    if (request.ApprovalStatus == ReleaseApprovalStatus.Approved
-                        && request.PublishMethod == PublishMethod.Scheduled
-                        && !request.PublishScheduledDate.HasValue)
-                    {
-                        return ValidationActionResult(ApprovedReleaseMustHavePublishScheduledDate);
-                    }
-
-                    var oldStatus = release.ApprovalStatus;
-
-                    release.ApprovalStatus = request.ApprovalStatus;
-                    release.NextReleaseDate = request.NextReleaseDate;
-                    release.PublishScheduled = request.PublishMethod == PublishMethod.Immediate &&
-                                               request.ApprovalStatus == ReleaseApprovalStatus.Approved
-                        ? DateTime.UtcNow
-                        : request.PublishScheduledDate;
-
-                    // NOTE: Subscribers should be notified if the release is approved and isn't amended,
-                    //       OR if the release is an amendment, is approved, and NotifySubscribers is true
-                    var notifySubscribers = request.ApprovalStatus == ReleaseApprovalStatus.Approved &&
-                        (!release.Amendment || (request.NotifySubscribers.HasValue && request.NotifySubscribers.Value));
-
-                    var releaseStatus = new ReleaseStatus
-                    {
-                        Release = release,
-                        InternalReleaseNote = request.LatestInternalReleaseNote,
-                        NotifySubscribers = notifySubscribers,
-                        ApprovalStatus = request.ApprovalStatus,
-                        Created = DateTime.UtcNow,
-                        CreatedById = _userService.GetUserId()
-                    };
-
-                    return await ValidateReleaseWithChecklist(release)
-                        .OnSuccessDo(() => RemoveUnusedImages(release.Id))
-                        .OnSuccess(async () =>
-                        {
-                            _context.Releases.Update(release);
-                            await _context.AddAsync(releaseStatus);
-                            await _context.SaveChangesAsync();
-
-                            // Only need to inform Publisher if changing release approval status to or from Approved
-                            if (oldStatus == ReleaseApprovalStatus.Approved || request.ApprovalStatus == ReleaseApprovalStatus.Approved)
-                            {
-                                await _publishingService.ReleaseChanged(
-                                    releaseId,
-                                    releaseStatus.Id,
-                                    request.PublishMethod == PublishMethod.Immediate
-                                );
-                            }
-
-                            _context.Update(release);
-                            await _context.SaveChangesAsync();
-
-                            if (request.ApprovalStatus == ReleaseApprovalStatus.Approved)
-                            {
-                                await _preReleaseUserService.SendPreReleaseUserInviteEmails(release);
-                            }
-
-                            return await GetRelease(releaseId);
-                        });
-                });
-        }
-
-        private async Task<Either<ActionResult, Unit>> ValidateReleaseWithChecklist(Release release)
-        {
-            if (release.ApprovalStatus != ReleaseApprovalStatus.Approved)
-            {
-                return Unit.Instance;
-            }
-
-            var errors = (await _releaseChecklistService.GetErrors(release))
-                .Select(error => error.Code)
-                .ToList();
-
-            if (!errors.Any())
-            {
-                return Unit.Instance;
-            }
-
-            return ValidationActionResult(errors);
         }
 
         public async Task<Either<ActionResult, TitleAndIdViewModel?>> GetLatestPublishedRelease(Guid publicationId)
@@ -587,31 +449,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
 
                             return await _releaseDataFileService.Delete(releaseId, fileId);
                         });
-                });
-        }
-
-        private async Task<Either<ActionResult, Unit>> RemoveUnusedImages(Guid releaseId)
-        {
-            return await _contentService.GetContentBlocks<HtmlBlock>(releaseId)
-                .OnSuccess(async contentBlocks =>
-                {
-                    var contentImageIds = contentBlocks.SelectMany(contentBlock =>
-                            HtmlImageUtil.GetReleaseImages(contentBlock.Body))
-                        .Distinct();
-
-                    var imageFiles = await _releaseFileRepository.GetByFileType(releaseId, Image);
-
-                    var unusedImages = imageFiles
-                        .Where(file => !contentImageIds.Contains(file.File.Id))
-                        .Select(file => file.File.Id)
-                        .ToList();
-
-                    if (unusedImages.Any())
-                    {
-                        return await _releaseFileService.Delete(releaseId, unusedImages);
-                    }
-
-                    return Unit.Instance;
                 });
         }
